+++ server/controllers/base.ts
@@ -47,7 +47,7 @@ class ControllerBase {
 			"/api/open/run_auto_test",
 			"/api/open/import_data",
 			"/api/interface/add",
-			"/api/interface/save",
+			"/api/interface/upsert",
 			"/api/interface/up",
 			"/api/interface/get",
 			"/api/interface/list",